textfilter: introduce newline indicator
Word2vec uses context to figure out the semantics of a word. The pandoc
filter serializes Text with formatting to plain text, discarding even
most of the line breaks. On the result, `split_whitespace()` is used to
only extract the words, no matter how much white space is in between. A
new line indicator has been introduced, which allows the Pandoc AST
filtering function to indicate, that a line break is required here. The
splitter function, which also removes punctuation, will change such a
byte to a proper line break.

This is now done for each paragraph-alike element, so that each
paragraph is a context.